+++ src/components/ChannelPreview.js
@@ -9,8 +9,8 @@ class ChannelPreview extends PureComponent {
     super(props);
 
     this.state = {
-      // unread: 0,
       lastMessage: {},
+      unread: 0,
       lastRead: new Date(),
     };
   }
@@ -20,6 +20,7 @@ class ChannelPreview extends PureComponent {
     activeChannel: PropTypes.object.isRequired,
     setActiveChannel: PropTypes.func.isRequired,
     Preview: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
+    watchers: PropTypes.object,
   };
 
   static defaultProps = {
@@ -29,6 +30,9 @@ class ChannelPreview extends PureComponent {
   componentDidMount() {
     // listen to change...
     const channel = this.props.channel;
+    const unread = channel.countUnread();
+
+    this.setState({ unread });
     channel.on('message.new', this.handleEvent);
   }