+++ src/components/ChannelPreviewLastMessage.js
@@ -26,7 +26,7 @@ class ChannelPreviewLastMessage extends PureComponent {
   channelPreviewButton = React.createRef();
 
   onSelectChannel = () => {
-    this.props.setActiveChannel(this.props.channel);
+    this.props.setActiveChannel(this.props.channel, this.props.watchers);
     this.channelPreviewButton.current.blur();
   };