+++ src/components/DateSeparator.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import { injectIntl } from 'react-intl';
 
 /**
- * DateSeparator - A simple date seperator
+ * DateSeparator - A simple date separator
  *
  * @example ./docs/DateSeparator.md
  * @extends PureComponent