+++ src/components/EditMessageForm.js
@@ -121,6 +121,7 @@ export class EditMessageForm extends React.Component {
           ref={this.props.emojiPickerRef}
         >
           <Picker
+            native
             emoji="point_up"
             title="Pick your emoji…"
             onSelect={this.props.onSelectEmoji}
@@ -154,6 +155,7 @@ export class EditMessageForm extends React.Component {
                 ref={this.props.emojiPickerRef}
               >
                 <Picker
+                  native
                   emoji="point_up"
                   title="Pick your emoji…"
                   onSelect={this.props.onSelectEmoji}