+++ src/components/MessageInputFlat.js
@@ -126,6 +126,7 @@ class MessageInputFlat extends PureComponent {
           ref={this.props.emojiPickerRef}
         >
           <Picker
+            native
             emoji="point_up"
             title="Pick your emoji…"
             onSelect={this.props.onSelectEmoji}
@@ -143,7 +144,7 @@ class MessageInputFlat extends PureComponent {
     return (
       <div
         className="str-chat__input-flat"
-        style={{ position: 'relative', zIndex: 100, width: '100%' }}
+        style={{ position: 'relative', zIndex: 1, width: '100%' }}
       >
         <ImageDropzone
           accept={this.props.acceptedFiles}