+++ src/components/MessageInputLarge.js
@@ -118,6 +118,7 @@ class MessageInputLarge extends PureComponent {
           ref={this.props.emojiPickerRef}
         >
           <Picker
+            native
             emoji="point_up"
             title="Pick your emoji…"
             onSelect={this.props.onSelectEmoji}