+++ src/components/MessageInputSmall.js
@@ -117,6 +117,7 @@ class MessageInputSmall extends PureComponent {
           ref={this.props.emojiPickerRef}
         >
           <Picker
+            native
             emoji="point_up"
             title="Pick your emoji…"
             onSelect={this.props.onSelectEmoji}