+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [0.6.27](https://github.com/GetStream/stream-chat-react/releases/tag/v0.6.27) 2019-09-20
+
+- Added `EmptyStateIndicator` prop to [ChannelList](https://getstream.github.io/stream-chat-react/#!/ChannelList) and [MessageList](https://getstream.github.io/stream-chat-react/#!/MessageList)
+- Added `watchers` prop to [ChannelList](https://getstream.github.io/stream-chat-react/#!/ChannelList) to specify [watchers pagination query](https://getstream.io/chat/docs/#channel_pagination) on `setActiveChannel`, including this makes one extra query on selecting a channel from the ChannelList
+- Updated react-images to version `1.0.0`
+
+## [0.6.26](https://github.com/GetStream/stream-chat-react/releases/tag/v0.6.26) 2019-09-10
+
+- Add IE 11 support for MessageInput
+- Fixing pagination issue when oldest message is not received yet
+- Fixing issue that didn't display unread count correctly on initial load
+
+## [0.6.25](https://github.com/GetStream/stream-chat-react/releases/tag/v0.6.25) 2019-09-05
+
+- The built in MessageInput components now use native emoji to create consistent rendering between the picker and the message
+
+## [0.6.22](https://github.com/GetStream/stream-chat-react/releases/tag/v0.6.22) 2019-08-15
+
+- Adding support for loading error indicator
+- Adding fallback as thumb_url for image attachments
+
 ## [0.6.21](https://github.com/GetStream/stream-chat-react/releases/tag/v0.6.21) 2019-08-05
 
 - Syncing and improvements in styleguide